// generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const ActiveStatus = {
  ACTIVE: 'ACTIVE',
  BLOCKED: 'BLOCKED'
} as const

export type ActiveStatus = (typeof ActiveStatus)[keyof typeof ActiveStatus]


export const Role = {
  USER: 'USER',
  AUTHOR: 'AUTHOR',
  ADMIN: 'ADMIN'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const PostStatus = {
  DRAFT: 'DRAFT',
  PUBLISHED: 'PUBLISHED',
  ARCHIVE: 'ARCHIVE'
} as const

export type PostStatus = (typeof PostStatus)[keyof typeof PostStatus]


export const CommentStatus = {
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED'
} as const

export type CommentStatus = (typeof CommentStatus)[keyof typeof CommentStatus]